+++ jd/generate-nutui-taro-vue.js
@@ -7,10 +7,10 @@ let importScssStr = `\n`;
 const packages = [];
 config.nav.map((item) => {
   item.packages.forEach((element) => {
-    let { name, show, type, taro, exportEmpty } = element;
+    let { name, show, type, taro, exportEmpty, exportEmptyTaro } = element;
     if (taro && (show || exportEmpty)) {
       importStr += `import ${name} from './__VUE/${name.toLowerCase()}/index${
-        exportEmpty ? '' : '.taro'
+        exportEmpty && !exportEmptyTaro ? '' : '.taro'
       }.vue';\n`;
       importScssStr += `import './__VUE/${name.toLowerCase()}/index.scss';\n`;
       packages.push(name);
@@ -33,24 +33,14 @@ const version = '${package.version}';
 export { install, version };
 export default { install, version};`;
 
-fs.outputFile(
-  path.resolve(__dirname, '../src/packages/nutui.taro.vue.build.ts'),
-  fileStrBuild,
-  'utf8',
-  (error) => {
-    // logger.success(`${package_config_path} 文件写入成功`);
-  }
-);
+fs.outputFile(path.resolve(__dirname, '../src/packages/nutui.taro.vue.build.ts'), fileStrBuild, 'utf8', (error) => {
+  // logger.success(`${package_config_path} 文件写入成功`);
+});
 let fileStrDev = `${importStr}
 ${installFunction}
 ${importScssStr}
 export { install, ${packages.join(',')}  };
 export default { install, version:'${package.version}'};`;
-fs.outputFile(
-  path.resolve(__dirname, '../src/packages/nutui.taro.vue.ts'),
-  fileStrDev,
-  'utf8',
-  (error) => {
-    // logger.success(`${package_config_path} 文件写入成功`);
-  }
-);
+fs.outputFile(path.resolve(__dirname, '../src/packages/nutui.taro.vue.ts'), fileStrDev, 'utf8', (error) => {
+  // logger.success(`${package_config_path} 文件写入成功`);
+});