+++ src/packages/__VUE/address/index.taro.vue
@@ -7,34 +7,29 @@
     v-model:visible="showPopup"
     :class="classes"
   >
-    <view-block class="nut-address">
-      <view-block class="nut-address__header">
-        <view-block class="arrow-back" @click="switchModule">
+    <view class="nut-address">
+      <view class="nut-address__header">
+        <view class="arrow-back" @click="switchModule">
           <nut-icon
             :name="backBtnIcon"
             color="#cccccc"
             v-if="privateType == 'custom' && type == 'exist' && backBtnIcon"
           ></nut-icon>
-        </view-block>
+        </view>
 
-        <view-block class="nut-address__header__title">
+        <view class="nut-address__header__title">
           {{ privateType == 'custom' ? customAddressTitle : existAddressTitle }}
-        </view-block>
+        </view>
 
-        <view-block class="arrow-close" @click="handClose('cross')">
-          <nut-icon
-            v-if="closeBtnIcon"
-            :name="closeBtnIcon"
-            color="#cccccc"
-            size="18px"
-          ></nut-icon>
-        </view-block>
-      </view-block>
+        <view class="arrow-close" @click="handClose('cross')">
+          <nut-icon v-if="closeBtnIcon" :name="closeBtnIcon" color="#cccccc" size="18px"></nut-icon>
+        </view>
+      </view>
 
       <!-- 请选择 -->
-      <view-block class="custom-address" v-if="privateType == 'custom'">
-        <view-block class="region-tab">
-          <view-block
+      <view class="custom-address" v-if="privateType == 'custom'">
+        <view class="region-tab">
+          <view
             class="tab-item"
             :class="[index == tabIndex ? 'active' : '', key]"
             v-for="(item, key, index) in selectedRegion"
@@ -43,24 +38,18 @@
             @click="changeRegionTab(item, key, index)"
           >
             <view>{{ getTabName(item, index) }}</view>
-          </view-block>
+          </view>
 
-          <view-block
-            class="region-tab-line"
-            ref="regionLine"
-            :style="{ left: lineDistance + 'px' }"
-          ></view-block>
-        </view-block>
+          <view class="region-tab-line" ref="regionLine" :style="{ left: lineDistance + 'px' }"></view>
+        </view>
 
-        <view-block class="region-con">
+        <view class="region-con">
           <ul class="region-group">
             <li
               v-for="(item, index) in regionList[tabName[tabIndex]]"
               :key="index"
               class="region-item"
-              :class="[
-                selectedRegion[tabName[tabIndex]].id == item.id ? 'active' : ''
-              ]"
+              :class="[selectedRegion[tabName[tabIndex]].id == item.id ? 'active' : '']"
               @click="nextAreaList(item)"
             >
               <nut-icon
@@ -74,10 +63,10 @@
               >{{ item.name }}
             </li>
           </ul>
-        </view-block>
-      </view-block>
+        </view>
+      </view>
 
-      <view-block class="custom-address" v-else-if="privateType === 'custom2'">
+      <view class="custom-address" v-else-if="privateType === 'custom2'">
         <view class="region-tab">
           <view
             class="tab-item"
@@ -89,11 +78,7 @@
           >
             <view>{{ getTabName(item, index) }}</view>
           </view>
-          <view
-            class="region-tab-line"
-            ref="regionLine"
-            :style="{ left: lineDistance + 'px' }"
-          ></view>
+          <view class="region-tab-line" ref="regionLine" :style="{ left: lineDistance + 'px' }"></view>
         </view>
         <view class="elevator-group" v-if="showPopup">
           <nut-elevator
@@ -102,10 +87,10 @@
             @click-item="handleElevatorItem"
           ></nut-elevator>
         </view>
-      </view-block>
+      </view>
 
       <!-- 配送至 -->
-      <view-block class="exist-address" v-else-if="privateType == 'exist'">
+      <view class="exist-address" v-else-if="privateType == 'exist'">
         <div class="exist-address-group">
           <ul class="exist-ul">
             <li
@@ -129,13 +114,7 @@
                 </div>
                 <div class="exist-item-info-bottom">
                   <view>
-                    {{
-                      item.provinceName +
-                      item.cityName +
-                      item.countyName +
-                      item.townName +
-                      item.addressDetail
-                    }}
+                    {{ item.provinceName + item.cityName + item.countyName + item.townName + item.addressDetail }}
                   </view>
                 </div>
               </div>
@@ -143,35 +122,28 @@
           </ul>
         </div>
 
-        <div
-          class="choose-other"
-          @click="switchModule"
-          v-if="isShowCustomAddress"
-        >
+        <div class="choose-other" @click="switchModule" v-if="isShowCustomAddress">
           <div class="btn">{{ customAndExistTitle }}</div>
         </div>
-      </view-block>
-    </view-block>
+      </view>
+    </view>
   </nut-popup>
 </template>
 <script lang="ts">
-import { reactive, ref, toRefs, watch, nextTick, computed } from 'vue';
+import { reactive, ref, toRefs, watch, computed } from 'vue';
 import { createComponent } from '../../utils/create';
 import Taro from '@tarojs/taro';
-import { transformData } from './transformData';
 
 const { create, componentName } = createComponent('address');
 
 interface RegionData {
+  id: string;
   name: string;
   [key: string]: any;
 }
-interface Region {
-  province: RegionData[];
-  city: RegionData[];
-  country: RegionData[];
-  town: RegionData[];
-  [key: string]: any;
+interface CustomRegionData {
+  title: string;
+  list: any[];
 }
 
 interface AddressList {
@@ -255,15 +227,7 @@ export default create({
       default: '200px'
     }
   },
-  emits: [
-    'update:visible',
-    'type',
-    'change',
-    'selected',
-    'close',
-    'close-mask',
-    'switch-module'
-  ],
+  emits: ['update:visible', 'type', 'change', 'selected', 'close', 'close-mask', 'switch-module'],
 
   setup(props, { emit }) {
     const classes = computed(() => {
@@ -288,10 +252,44 @@ export default create({
 
     const isCustom2 = computed(() => props.type === 'custom2');
 
+    const transformData = (data: RegionData[]) => {
+      if (!Array.isArray(data)) throw new TypeError('params muse be array.');
+
+      if (!data.length) return [];
+
+      data.forEach((item: RegionData) => {
+        if (!item.title) {
+          console.error('[NutUI] <Address> 请检查数组选项的 title 值是否有设置 ,title 为必填项 .');
+          return;
+        }
+      });
+
+      const newData: CustomRegionData[] = [];
+
+      data = data.sort((a: RegionData, b: RegionData) => {
+        return a.title.localeCompare(b.title);
+      });
+
+      data.forEach((item: RegionData) => {
+        const index = newData.findIndex((value: CustomRegionData) => value.title === item.title);
+        if (index <= -1) {
+          newData.push({
+            title: item.title,
+            list: [].concat(item)
+          });
+        } else {
+          newData[index] = {
+            title: item.title,
+            list: newData[index].list.concat(item)
+          };
+        }
+      });
+
+      return newData;
+    };
+
     const regionList = reactive({
-      province: isCustom2.value
-        ? transformData(props.province)
-        : props.province,
+      province: isCustom2.value ? transformData(props.province) : props.province,
       city: isCustom2.value ? transformData(props.city) : props.city,
       country: isCustom2.value ? transformData(props.country) : props.country,
       town: isCustom2.value ? transformData(props.town) : props.town