+++ src/packages/__VUE/popover/index.vue
@@ -1,30 +1,28 @@
 <template>
-  <div @click="openPopover()" :class="classes">
-    <slot name="reference"></slot>
-
+  <view @click="openPopover" :class="classes">
+    <div ref="reference"> <slot name="reference"></slot></div>
     <template v-if="showPopup">
-      <div class="more-background" @click="closePopover()"> </div>
-      <div :class="popoverContent">
-        <div :class="popoverArrow"> </div>
+      <view class="more-background" @click="closePopover"> </view>
+      <view :class="popoverContent" :style="getStyle">
+        <view :class="popoverArrow" :style="getArrowStyle"> </view>
 
         <slot name="content"></slot>
 
-        <div
-          v-for="item in iconItemList"
+        <view
+          v-for="item in list"
           :key="item.name"
           :class="{ 'title-item': true, disabled: item.disabled }"
+          @click="chooseItem(e, item)"
         >
-          <slot v-if="item.icon">
-            <nut-icon class="item-img" :name="item.icon"></nut-icon
-          ></slot>
-          <div class="title-name">{{ item.name }}</div>
-        </div>
-      </div>
+          <slot v-if="item.icon"> <nut-icon class="item-img" :name="item.icon"></nut-icon></slot>
+          <view class="title-name">{{ item.name }}</view>
+        </view>
+      </view>
     </template>
-  </div>
+  </view>
 </template>
 <script lang="ts">
-import { onMounted, computed, watch, ref, PropType, toRefs } from 'vue';
+import { onMounted, computed, watch, ref, PropType, toRefs, reactive, CSSProperties } from 'vue';
 import { createComponent } from '../../utils/create';
 const { componentName, create } = createComponent('popover');
 import Popup, { popupProps } from '../popup/index.vue';
@@ -42,7 +40,7 @@ export default create({
   },
   props: {
     ...popupProps,
-    iconItemList: {
+    list: {
       type: Array,
       default: []
     },
@@ -57,8 +55,13 @@ export default create({
       default: 'bottom'
     }
   },
-  emits: ['update', 'update:visible', 'close'],
+  emits: ['update', 'update:visible', 'close', 'choose', 'openPopover'],
   setup(props, { emit }) {
+    const reference = ref();
+    const state = reactive({
+      elWidth: 0,
+      elHeight: 0
+    });
     const showPopup = ref(props.visible);
 
     const { theme, location } = toRefs(props);
@@ -87,7 +90,53 @@ export default create({
       };
     });
 
-    onMounted(() => {});
+    function getReference() {
+      const domElem = document.documentElement;
+      state.elWidth = reference.value.offsetWidth;
+      state.elHeight = reference.value.offsetHeight;
+    }
+
+    const getStyle = computed(() => {
+      const style: CSSProperties = {};
+      if (location.value == 'top') {
+        style.bottom = state.elHeight + 20 + 'px';
+        style.left = 0 + 'px';
+      } else if (location.value == 'right') {
+        style.top = 0 + 'px';
+        style.right = -state.elWidth + 'px';
+      } else if (location.value == 'left') {
+        style.top = 0 + 'px';
+        style.left = -state.elWidth + 'px';
+      } else {
+        style.top = state.elHeight + 20 + 'px';
+        style.left = 0 + 'px';
+      }
+
+      return style;
+    });
+
+    const getArrowStyle = computed(() => {
+      const style: CSSProperties = {};
+      if (location.value == 'top') {
+        style.bottom = -20 + 'px';
+        style.left = state.elWidth / 2 + 'px';
+      } else if (location.value == 'right') {
+        style.top = 20 + 'px';
+        style.left = -20 + 'px';
+      } else if (location.value == 'left') {
+        style.top = 20 + 'px';
+        style.right = -20 + 'px';
+      } else {
+        style.left = state.elWidth / 2 + 'px';
+        style.top = -20 + 'px';
+      }
+
+      return style;
+    });
+
+    onMounted(() => {
+      getReference();
+    });
 
     watch(
       () => props.visible,
@@ -101,14 +150,24 @@ export default create({
       emit('update:visible', val);
     };
 
-    const openPopover = () => {
+    const openPopover = (event: Event) => {
+      event.stopPropagation();
+      event.preventDefault();
       update(!props.visible);
+      emit('open');
     };
 
-    const closePopover = (e) => {
-      e.stopPropagation();
+    const closePopover = (event: Event) => {
+      event.stopPropagation();
+      event.preventDefault();
       emit('close');
       emit('update:visible', false);
+    };
+
+    const chooseItem = (event: Event, item: any) => {
+      event.stopPropagation();
+      event.preventDefault();
+      emit('choose');
     };
 
     return {
@@ -117,7 +176,12 @@ export default create({
       openPopover,
       popoverContent,
       popoverArrow,
-      closePopover
+      closePopover,
+      chooseItem,
+      getReference,
+      reference,
+      getStyle,
+      getArrowStyle
     };
   }
 });