+++ vite.config.build.taro.vue.disperse.ts
@@ -14,10 +14,10 @@ let input = {};
 
 configPkg.nav.map((item) => {
   item.packages.forEach((element) => {
-    let { name, show, taro, type, exportEmpty } = element;
+    let { name, show, taro, type, exportEmpty, exportEmptyTaro } = element;
     if (taro && (show || exportEmpty)) {
       input[name] = `./src/packages/__VUE/${name.toLowerCase()}/index${
-        exportEmpty ? '.vue' : '.taro.vue'
+        exportEmpty && !exportEmptyTaro ? '.vue' : '.taro.vue'
       }`;
     }
   });
@@ -44,7 +44,7 @@ export default defineConfig({
     },
     rollupOptions: {
       // 请确保外部化那些你的库中不需要的依赖
-      external: ['vue', 'vue-router', '@tarojs/taro', 'pinyin'],
+      external: ['vue', 'vue-router', '@tarojs/taro'],
       input,
       output: {
         banner,